[kernel][arm64] Use adr_global for __code_start in _start
Plain adr has a distance limit.  In an unoptimized build with the object
files reordered, the _start code might not come first in the link and so
could be far away from __code_start.  It shouldn't matter how the code is
laid out.  So use adr_global to handle the general case here.

Bug: BLD-325
Test: CQ & manual build with GN and -O0
Change-Id: I19321e6091f725e6f45842f05eb4db4b1910e74d