[debugger] Improve inline stepping.
Properly handle "step over" for inline frames.

Fixes an off-by-one error in ThreadController::SetInlineFrameIfAmbiguous
where it would not consider non-inline frames as the frame to set to.
This condition should be that everything above the frame being
considered is an ambiguous inline location, but not necessarily the
frame itself.

Reset the hide_ambiguous_frame_count in the Stack object whenever the
stack is changed. Inheriting this value wehn the stack is changed is
incorrect.

When ThreadController::SetInlineFrameIfAmbiguous fails to find the
requested frame, it will now put back the value to the old one rather
than reset it to 0. This is important when there are nested controllers
(in the case of "step over" this can clobber what the sub-"finish"
controller did).

The step controller enhancements require file/line information which
requires that the thread controller tests have a lot more knowledge
about line numbers. The test harness was enanced to support this, and
some helper functions for constructing line details were enhanced.

Change-Id: I0290c4fccc09dd4c2d8ff32344ca4eff190a05af