[aml][rawnand] Reduce the cmd fifo timeout in the read path to 1 second.
Trying to track down where reads could get blocked/spin in the amlogic
rawnand driver. The only culprit is the 1048 second timeout in the read
path, waiting for the cmd FIFO to drain. There is no reason this has to
be greater than 1 second. If this is the culprit, we should see errors
logged, the read cmd retried by the NAND layer etc.

Bug: ZX-3408

Test: Flashing and doing basic NAND testing.
Change-Id: Ifbff43f97e46cd84ac188f390dbc6ed408c223c6