[scenic][SCN-1054] Add SetTranslationRH stubs to facilitate rollback of handedness change.
By itself this change is a no-op. The intention is to land this change
underneath fxr/250252 so that if fxr/250252 needs to be rolled back after
dependent changes (e.g. fxr/247814) have landed it can be done so safely.

This will also allow petal changes to land before the fuchsia.git change which
may be desireable as it will reduce exposure to SCN-1054 deprecation warnings.

Change-Id: Idc2e44caf6d727e802a1ec6f0caa3e12678fcee2