[carnelian] Adopt euclid for geometry
Euclid is a very nice crate for dealing with 2D and 3D geometry.

Fixed voila’s use of carnelian, even though it is not currently in the build. I added
in temporarily for testing but don’t want to add it permanently until the ledger
team concurs.

Also set up .cargo/config for fuchsia.git and removed it from garnet.

Change-Id: I33246b1e0bbe8608fdb8c5a3d6ea89978d4dae44