[fidl][difl] Start of tool that evalualtes FIDL changes
This is a framework for a tool that compares two FIDL libraries and
lists the potentially API and ABI breaking changes. FIDL libraries are
consumed as FIDL IR JSON files as produced by the fidlc compiler.

The tool is extremely incomplete in the following ways:
 - it only checks a single library, not a set of many libraries
 - it doesn't identify all of the changes between libraries
 - it only classifies ABI changes not API changes
 - its output isn't actionable
 - its output is ugly
 - its output can't be consumed by tools like Tricium

There's the start of a test framework in particular two versions of a
library under garnet/public/lib/fidl/tools/difl_test_fidl/ which is
integrated into the build. It uses specially formatted comments in the
FIDL source to track which changes are expected to be detected. It is
incomplete:
 - some changes the tool identifies aren't in the test library
 - some changes in the test library aren't identified by the tool
 - only the types of the changes, are checked, not any additional
   metadata

The tool can be run like:
  ./scripts/difl \
    --old out/x64/gen/garnet/public/lib/fidl/tools/difl_test_fidl/before.fidl.json \
    --new out/x64/gen/garnet/public/lib/fidl/tools/difl_test_fidl/after.fidl.json

Change-Id: I376b8aa2266ef8353ee8f70d9605f59257d9fb04