[debugger] Set containing code for inline functions.
There are two types of parents that an inlined function can have: the
lexical parent (e.g. containing namespace or class of that function
declaration), and the parent the code is inlined into.

Previously the parent() would return only the lexical parent which makes
it impossible to walk up the inlined function hierarchy.

This change adds a separate concept for the containing code block for
inline functions for traversing this data.

Change-Id: I14fbb6f604e33a972e276e2da5bab02ac8421d87