[wlan] Use explicit "LATE" timer slack rather than implicit "CENTER"
Passing 0 for the options parameter of zx::timer::create yields a
timer with ZX_TIMER_SLACK_CENTER meaning that the timer may fire
before or after its deadline (up to the amount of slack passed to
zx::timer::set).

Making the slack type explicit because in a future CL, we may change
zx::timer::create to interpret 0 as "use the default slack type".

Change-Id: I6dc2c7dd974da3634fddbffb77b9a267cdfe0667