[bt][data] limit the size of the socket-write-queue
If a profile stops reading from the zx::socket it receives in the
Profile.OnConnected event, we will start queueing inbound frames
for that zx::socket. First, we queue frames into the zx::socket,
and then into a queue within SocketChannelRelay.

Given the relatively large size of the zx::socket queue [*], it is not
actually clear that we need any queuing within SocketChannelRelay
itself. And, of course, an unbounded queue (as we currently have) will
eventually lead to an out-of-memory condition.

For now, we cap SocketChannelRelay's internal queue at two frames. We'll
revisit the question of completely removing the queue at another time.

[*] The zx::socket's queue is large enough for 1 full second of
inbound data, assuming an input rate of 2 Mbps.

Bug: BT-116
Test: bt-host-unittests (on eve)
Change-Id: I2f13371b90c4db5b6ee70565c839e017c4d4a867