[ethernet] Handle multiple opens on an instance
This can happen if fuchsia.io.Node:Clone() is invoked on an open
instance connection.  Previously the driver would get confused and the
first close() would invalidate the state.  This patch makes it so the
state will only be invalidated after all of the close()'s are invoked.

ZX-3459 #comment Improve behavior of the ethernet driver

Change-Id: I032dea5612e2ce8e2685079a4ba2d0400a1e0344