[dev][bus][platform] Report a pdev instance for cpu_trace
This moves the code from board/x86 to the bus/platform code so that it is
used by arm as well. The purpose of this code is to let us route a
BTI to cpu_trace so it can lookup physical addresses.

Also, change device VID to GENERIC in preparation for arm port.
The "device" is a virtual one, essentially independent of chip/vendor.

Tested: CQ
fx shell cpuperf --spec-file=/pkgfs/packages/cpuperf/0/data/basic-cpu-memory.cpspec

Change-Id: Icc3170f66b0b32b020a101fc6f07bd024eec1683