[rust][async] Add control plane support to Socket
This rewrites Socket to not use RWHandle, and implements reading and
writing from the control plane.

This is done by having a generic socket abstraction, which implements
AsyncRead, AsyncWrite, Stream, etc, and can be used to represent either
the control plane or the data plane of the socket, depending on which
PacketReceiver is used.

Change-Id: Ia2005dcb45ed0dc16b8341b5faf87974079ea13a