[zircon] Replace ZX_OBJ_TYPE_LAST with ZX_OBJ_TYPE_UPPER_BOUND
System API rubric says not to have a "last" or "count" declaration for
enumerated values because we might need to expand the list of enumerated
values in the future.

However, for object type, we already commit to a maximum number of
object types as part of zx_info_process_handle_stats_t. This CL switches
the ZX_OBJ_TYPE_LAST to ZX_OBJ_TYPE_UPPER_BOUND and matches that value.

"64 types of kernel objects ought to be enough for anybody."
  -- Adam Barth, 2019

Change-Id: I7715b46459df165358dd935633c970471eb0621f