[zircon][utest][pager] Exclude root-resource tests at runtime, not compile
Rather than using #ifdef BUILD_COMBINED_TESTS to decide whether the
root resource handle will be available for tests that need it, instead
use a weak symbol for get_root_resource and check it at runtime.
This obviates the need to compile the test sourcce file differently for
the integrated and standalone cases.

Bug: BLD-325
Test: CQ
Change-Id: I48e3f192af87da354f05cd9b75c6e1124d4e2106