[zircon] [gn] Convert convenience scripts to use GN build
The obsolete scripts build-all-zircon, make-parallel, and
make-release are gone entirely.  The build-zircon script is
updated to drive the GN+Ninja build using approximately the
same command line UI.  The buildall script is updated to use
build-zircon.  The run-zircon script is updated to default to
the new locations used by build-zircon.

Bug: BLD-325
Test: ./scripts/buildall -c -A -r && ./scripts/run-zircon-x64
Change-Id: I4c8b71a804ef81b71bd535e6bdef28e683f22450

Bug: BLD-325
Change-Id: If88c836c805ab92657f091371c6a4033b37aa281