[componentmgr] Model tracks children
This change turns Model into a hierarchy of realms which represent the
component instance tree. Children are resolved when the component
instance is resolved, which happens either when someone binds to the
instance and causes it to start or when it needs to be resolved because
it is a descendant of another unresolved component instance.

CF-582 #done

TEST=bind_instance unit tests

Change-Id: I93efff161efd0c61b6be05e337bb122f0eb1d927