[tap] Preserve whitespace in strings
This change causes the parser to output TAP documents which have
preserved the whitespace for test line descriptions and directives.

Note that the whitespace surrounding a description or directive is
assumed to be unimportant and is not preserved.  For example: The
text "some  text" is parsed as "some   text", but we assume
that it's OK to trim " some text " when parsing, as "some text".

Change-Id: Ia16a2ae8b1c97e99533c5768a0a08e78d21aa3d7