[sysmem] Add CheckBuffersAllocated
The display driver wants to be able to import buffers from a buffer
collection without blocking indefinitely (while still using the C
bindings). The new CheckBuffersAllocated method can be used to ensure
that the initiator waited for the buffers to be allocated before trying
to do the import, guaranteeing that WaitForBuffersAllocated will return
quickly if the sysmem driver is behaving properly.

WaitForBuffersAllocated could be modified to have an argument saying
whether it should block or not, but the component that checks whether
the buffers are allocated (the display core driver) doesn't need to get
the details of the actual allocated buffers.

Test: sysmem-tests on VIM2

Change-Id: I83095ec2c1c7f5eeaba951651911ceaf845498bc