[story_storage] Fix use-after-free when writing to story storage
When OnPageChange is called on story storage it will resolve any futures
which have been added by write operations which are waiting for
confirmation that the write completed through the ledger.

Completing such a future may trigger a chain of events which eventually
ends up deleting the story storage instance (e.g.
DispatchStoryCommandExecutor will tear down the story storage associated
with a particular set of commands when they complete).

Thus completing the waiting-for-write-futures must be the last thing to
happen in the OnPageChange implementation.

MF-193 #comment Fix OnPageChange ASan issue.

Change-Id: I663c9941dcb0e5b865f6f25b7432f6ca9d6da836