[bt][l2cap] remove some mistaken static_asserts
Asserting that the various l2cap_internal structs are standard-layout
was a neat idea, but a) I botched the implementation, and b) after
fixing the bugs, I realized that standard-layout isn't quite what
I'm looking for. So let's remove the assertions --- the unit tests
verify the behavior more precisely anyway.

In more detail: my patch repeatedly asserted that the same struct was
standard layout, instead of checking each struct for that property.
But it turns out that SimpleStartOfSduFrameHeader _isn't_
standard-layout, and it shouldn't be. For that struct to be
standard-layout, |sdu_len| would have to be at offset 0. But that
would be at odds with the wire format.

Bug: None
Test: bt-host-unittests (on eve)

Change-Id: Ia4193f0b5f5ab8411024d5cf0e6764d25a73ff1c