[user_provider] Fix use-after-free of user controller
Previously user controllers were deleted in two places, once in the
callback to UserController::Logout, and once in the callback which is
passed to the user controller's constructor.

The problem was that the the logout callback would be called, the user
controller would get destructed, and then control flow would return to
the session manager's teardown callback, which again accessed the user
controller (to call the callback passed to its constructor).

Now only the final callback deletes the user controller, which is still
not very clean but is at least symmetrical since the callback is passed
in to the user controller constructor.

MF-193 #comment Fix user provider ASan issue

Change-Id: Ic1905b0236947ec6680006343e4ec80f4e8b1d92