[sysmem] keep constraints beyond clean Close()
For some participants, it's convenient to be able to SetConstraints() on
the BufferCollection view and then immediately do a clean Close() on the
BufferCollection.  With this change, those constraints will be retained.

It's still required that at least one participant or the initiator keep
their BufferCollection view alive for the LogicalBufferCollection to
remain allocated.

A test is added to check that constraints are kept from a participant
that does a clean Close().

Tested: sysmem-test, sysmem_connector_test
Change-Id: If2ad845cce62d31cfc059470f033ff1f6e842e0f