[build][cmc] Implement build-time whitelist mechanism for deprecated-global-persistent-storage cmx feature
This adds a new 'extra_schemas' argument to cmc's 'validate' command
that can be used to provide extra JSON schemas to validate component
manifests against. This is used to implement build-time validation for
the 'deprecated-global-persistent-storage' feature so that only
whitelisted targets are allowed to use it.

The whitelist is not actually populated here. It will be populated in a
subsequent change that makes the other the mechanical changes to the
manifests and BUILD.gn files.

CF-28 #progress

Test: fx run-host-tests cmc_host_tests
Test: Validated whitelist mechanism by adding
"deprecated-global-persistent-storage" to a .cmx file and verified build
failed as expected.

Change-Id: Ieefe4a9e745d0680f82d5bbf5e815a203acc8e96