[zircon][unittest] Use monotonic clock for watchdog
CLOCK_REALTIME (i.e. ZX_CLOCK_UTC) is not a good clock to use for
measuring intervals. In particular, it jumps forward significantly at
boot when it is initialized with zx_clock_adjust. If this occurs while
the watchdog is blocked, using REALTIME will result in a much larger
calculated elapsed time that was actually the case.

Test: Add indefinite blocking to test, observe that watchdog fires.
Test: Add 500ms delay in sys_clock_adjust, run futex-test with autorun,
observe TestFutexWaitTimeoutElapsed fails w/REALTIME but not MONOTONIC.
ZX-3493 #done [zircon][unittest] Use monotonic clock for watchdog

Change-Id: Ibc4f43c0ac7c7dda4d52e52711b8c91cc21d63fc