[scenic] Create View and ViewHolder Nodes
Updated the implementation of View to be a View + ViewNode pair,
and the ViewHolder to a ViewHolderNode.

This cleans up the View/ViewHolder management between the ResourceMap
and the scene graph. ViewHolders are treated as nodes because they have
the same lifecycle as a node w.r.t the scene graph and the ResourceMap.
The ViewHolderNode functinality is limited such that a ViewHolderNode
can only parent the linked View's ViewNode. The public View is implemented
as a View that manages a ViewNode to support existing behavior that the
ResourceMap manages the lifetime of a View. The ViewNode is the View's
explicit link into the scene graph. The transforms in the graph above
the ViewNode define the View's basis.

SCN-1176#done
SCN-130#done

SCN-820#comment

Tested:
ran on astro
added unittests
dumped scenes

Change-Id: I4c0be44284615c60e7566a6cff1576991ed67da2