[debugger] Improve stepping to inlines.
Changes the step controller to always set the ambiguous inline location
to the oldest frame no matter how we got there. This fixes a problem
when doing "step" to step out of a function call that lands immediately
in a call to an inline function. Previously the user would see the first
instruction of the inline, and how the user will see the inline call.

The code that handled this case for "step over" can now be removed as
this is a more general solution to the same problem. The existing step
over test mostly covers this test case.

The existing test was incorrect as the ranges didn't match the desired
testing case. The new step over test now implicitly tests the changed
code.

Change-Id: I1e898c9e213efde2e7d5d492cab0019afc599b5f