[bt][tests] Re-organise Bluetooth Integration tests
This CL performs some housekeeping of the bluetooth integration tests.
The host_integration and host_lifecycle binaries are combined into a
single binary, which is reorganised to match the standard Rust binary
layout (with source code in src/).

The code is further split up into meaningful modules, with the tests
themselves in `crate::tests`, and test harness code in `crate::harness`.

The test harness for host tests has been reorganised, with `HostTestPtr`
renamed to `HostTest` as it is the public API for host tests, and it
provides a simple but powerful expectation API. The previous
`HostTest`, has been renamed to `HostTestInner` and is a private
implementation detail of the `crate::harness::host_driver` module, and
is not exported outside that module.

The final binary now runs all tests, reporting all failures but continuing on
rather than aborting at the first failure.

The previous structs used to compare for expected states have been
replaced with a boolean Predicate system, which is defined in the
fuchsia_bluetooth library. This provides a compositional module based on
boolean algebra, which allows arbitrary tests to be easily set up,
whilst also providing useful primitives such as address and name
comparison to be easily combined in a literate manner.

test: The new tests can be run as a single binary:
  (after fx set --available=garnet/packages/tests/bluetooth)
  $ fx build-push bluetooth-tests
  $ fx shell /pkgfs/packages/bluetooth-tests/0/test/bt-host-integration-tests

  The tests for the expectation predicates can be run by running the
  tests for the fuchsia bluetooth crate:
  $ fx run-test bluetooth-crate-tests

Change-Id: I5b3835ef20fa7bd9c9305139cfe45a2302ae7649