[kernel] Avoid rescheduling when waking reader-writer lock threads
The correctness of the reader-writer lock relies on completely executing the sections
where the thread_lock is held. If reschedules can happen then the lock can be left
in an inconsistent state and lead to lock violations. This change just disables
rescheduling for the duration of any lock portions where we hold the thread lock
and wake threads.

ZX-3494

Test: Ran kernel unit tests
Change-Id: Ibba600f387366f874bea59555e8ef5b13de2490a