[zircon] Unify Dispatcher's canaries
Before this CL canaries were ad-hoc placed in final derived classes
with this change now all canaries are in a single place and assert()
methods can be done at the SoloDispatcher and PeeredDispatcher
level.

Sadly, this cannot be simply extended to the root of the hiearchy
because the template parameter (magic) infects the Dispatcher type.

That is not to say there are no options, but this is big enough
for this CL.

TEST = existing tests suffice. The canary tests that memory is
properly used (and not stomped) in a transparent way.

- Included in this change I made the Dispatcher ctor protected.
- Some magic strings have changed, this should be a no-op

ZX-3472 #comment progress

Change-Id: I4bd20f1d19a290dfdb40ef2bfeeb8818e89be1cd