[runtests] stop trying to sync the filesystem
The code here appears to be trying to force a system to write data to disk,
but it appears also to have absolutely no reason to be performing this
operation. Test results are read back from a live system. As there are no
comments explaining the original intent, that the implementation isn't having
any observable side-effect (literally does nothing but error on Fuchsia, and
doesn't have any significant meaning on host other than being slow), it is
best removed entirely.

Change-Id: I590e15cdf39b2de24528e7ccaa181bcf210da7bb