[scenic] Add hello_touch_colors example app
Add a new example app, hello_touch_colors.  hello_touch_colors is
essentially the part of hello_base_view that changes the color of the
screen when the screen is touched (so in particular, it does not have
the shader toy view in the background that continuously renders).  It
also contains additional trace events.

The app is intended to be used for input latency benchmarking.

Change-Id: Iaaca6cd9a91d5d8237466a85da983c3ced7dcc36