[products] Add an sdk image product configuration.
This product configuration is intended to be used for images produced
along with the SDK.
The first version is a carbon-copy of the packages that were included
in the current SDK images. Future versions will move most of the
packages into the `available` set instead of the `monolith`, producing
an very basic system image booting to a shell.

Bug: DX-1034
Test: `fx set x64 --packages topaz/packages/sdk/topaz` and `fx set x64
      --product sdk_image` produce the same system_index#monolith set.

Change-Id: I26ff9cf907256d54e2ce714be744588e116dad3a