[kernel][vm] Add VmPageList foreach variant with gaps
There are multiple places in VmObjectPaged where it iterates over the
page list and needs to handle missing pages in the range. This change
refactors the duplicated logic into a ForEveryPageAndGapInRange method
on VmPageList that invokes a callback for every page in the range as
well as for every run of missing pages in the range.

Test: k ut vmpl
Change-Id: Icf13a08a46cb1114c3e43f13fae4f163015140f2