[ldmsg] Prepare to fix ldmsg reply lengths
Previously, we would generate a 20 byte reply for several methods that
should have had a 24 byte response. This CL loosens the validation in
receivers so that they will accept either a 20 byte or 24 byte response.

Once all the prebuilt programs update to include this change, a later CL
will update the senders to actually send 24 bytes.

DX-1008 #comment

Test: ldsvc_tests
Change-Id: Icc737ce6af96533d718d7dc1dad8ec2d5fa0f97e