[debugger] Improve inline stepping.
Part 1: Previously "step into inline" functionality was in a separate
"step into" thread controller since this doesn't actually run the
thread. But for consistency and general sanity other controllers (like
the "step over" one) need to see these fake stops so they can do the
right thing in response to the changing stack.

This moves all inline handling into the existing "step" controller which
now handles all variants of stepping into. The existing test is moved
and updated.

Part 2: The step controller keeps the file/line information when
stepping by line so it can see if the line is the same even if all of
the ranges changed. This breaks when the file line the user is on isn't
in the line table, which happens when there's an inline call by itself
on a line. In this case, the stack will synthesize a stack frame for the
call to the inline frame at the call location, even though that line
doesn't exist in the line table.

The update is to always use the file/line from the stack which allows
proper stepping into inline frames. This allows stepping over code like:

  NonInlineFunction(inline_function());

And:

  InlineOne(); InlineTwo();

Change-Id: I7fbbda6a67152651a4aa78623ff545a16a9e82ee