[persistent-storage] Rename feature to 'deprecated-global-persistent-storage'
This modifies all component manifests in this repo that previously
included the 'persistent-storage' sandbox feature to instead include
'deprecated-global-persistent-storage', which is the new name for the
same behavior. There is no functional impact of this change.

This also adds all migrated components to the build-time whitelist that
guards usage of 'deprecated-global-persistent-storage'. Only targets in
that whitelist are allowed to use this feature since it is deprecated.
The whitelist should not be added to after this initial rename.

CF-28 #progress

Test: CQ, aka successful build (whitelist wasn't tripped) and no
regressions in behavior.

Change-Id: Ib4c07ba90efa5f20719d4d444cc8c212f210e1ec