[zircon][unittest] Use realtime with cond
pthread conds use REALTIME by default. It's possible to change this
with condattr, but OSX doesn't define the necessary functions. So just
use the realtime clock for the timed wait. This can result in a
spurious wakeup when realtime is initialized, but that's okay.

Test: observe exception-test runs quickly
ZX-3510 #done [zircon][unittest] Use realtime with cond

Change-Id: If2f60ae2b7d8e1843e988bc7607b80826f41ce1a