[tracing][msd-intel-gen] connect command buffer submission to end of Context Exec
Trace events were not showing up when there were multiple Magma clients,
since it was possible that another Context Exec was running when any
given command buffer was submitted. Connecting the flow to the end of
the Context Exec fixed this, because we were guaranteed that the correct
Context Exec would be running.

Test: tracing with multiple msd clients, such as hello_base_view

Change-Id: I251c26bf6f6cebe3729f907cf0ceee55f126fc28