[ledger] Update (Has|Get)Object API in PageDb.
When PageDb::GetObject was called with a nullptr as object, it had
the same result and performance as HasObject. When reading code, this
API was confusing, as it was sometimes hard to understand if we needed
the contents of the object or not. When writing code, it was prone to
accidently passing a non-null object, thus getting the object contents.

This CL removes this behavior from GetObject, and now it should only be
used when we need the contents of the object.

Also, HasObject no longer has a bool argumen, but rather uses the
returned status for the result.

Change-Id: I8dbe9a29f7ba830278647ac04509300c89eedce0