[docs][fidl] Updating 'interface' -> 'protocol' keyword
Test: Read the docs

This change is intended as 1 of 2; in this first pass, we change the
keyword `interface` to the keyword `protocol`. Only the bare minimum
changes to support this are done; a much larger set of changes, that
examines the word "interface" and ensures the context, is planned
for part 2.

Minor column width and other fixup edits are performed herein as well.

Change-Id: Id7e8b470410b3cc024fa174d4f954af3035896ab