[cobalt] Start using ProjectContextFactory
CB-220

This is part of a series of CLs whose goal is to clean up the client-side
infrastrcutrue around accessing the Cobalt Registry.

- Start using the new ProjectContextFactory. That class encapsulates all of the
logic around differentiating between the different cases for the state
that a CobaltRegistry might be in:
  (a) Containing a single Cobalt 0.1 project
  (b) Containing a single Cobalt 1.0 project
  (c) Containing multiple projects.
This allows us to eliminate all references to the lower-layer config-related
objects such as ClientConfig and ProjectConfigs.

- Start using the new constructor for Logger::Logger() that has the
Logger own its ProjectContext. This eliminates the need to hold on
to a ProjectContext at this layer.

- Get rid of extra instance of CobaltRegistry for the internal metrics
project. This is not necessary since that project is contained in the
global metrics registry.

- Clean up some code duplication in logger_factory_impl.cc.

Change-Id: I08bfcb60544bca6850af8406102f69256a4b25da