[bt][l2cap] constify bits of channel
There are some fields of Channel that are const in practice, but not
declared as such. Update the declaration to reflect their constness.
And to make this possible: remove an (unused) setter in FakeChannel,
which could violate constness.

The rationale for wanting these to be |const| is so that we can safely
use them to initialize fields of the (soon-to-be-added) transmit
engine. That engine will be constructed in the Channel constructor,
and it (the engine) will cache the values of some Channel parameters.

Note that the transmit engine only needs the |id_| and |tx_mtu_|
fields to be const. But, while I'm here, it seems like good hygiene to
mark the other fields const as well. Semantically, most of these fields
shouldn't change during the life of a Channel.

The MTU fields are different, in that the protocol allows for these to
be changed dynamically. But let's cross that bridge when we come to
it.

Bug: BT-442
Bug: BT-532
Test: bt-host-unittests (on eve)
Change-Id: I32a91d1a569970c5ae7e47f32d43c57b69c66318