[bt][l2cap] add TxEngine abstraction
Refactor the Channel code, to separate out accepting an SDU for
transmission from actually transmitting the resulting PDUs. This
separation is necessary to support Enhanced Retransmission Mode
(ERTM), which needs to a) queue packets when the receiver is busy, and
b) retransmit packets if they are not acknolwedged within a timeout.

Note that this CL does not add any support for ERTM itself.  That will
come in later CLs.

Bug: BT-532
Test: bt-host-unittests (on eve)

Change-Id: Iac6b9e5104f75f259b68633648e6ea5a5f806c54