[devmgr] Move kernel-debug logic to service.
We create a standalone(ish) service that lives in svchost that will
handle kernel-debug related things (kerneldebug commands and ktrace).
This also update the "dm" tool to talk directly to this new service.

This is a re-commit of I5afd3e7bc366debf7b4819dfe62d66cfc665a9dc with
support for the "k" cmdline tool.

Test: Tested 'dm' tool and 'runtests' on qemu-x86.
Issue: ZX-3403
Change-Id: I5121346af4d24d479872c8655258b8cf4abb9957