[bt][common] Force DeviceAddress to be 8 bytes
The 'Type' enum class in common::DeviceAddress has only 4 options, but
defaults to a 32bit int, which due to alignment causes DeviceAddress to
be 12 bytes. Switching to an explicit uint16_t allows the entire class
to be packed into 8 bytes, which will allow it to fit in a 64-bit
register and is likely to generally improve memory usage and performance.

Change-Id: I6b1af5bde495bc57ee41db4f6b7e664d056f24cc