[kernel][arm][gic] disable the gic v2 vs v3 detection mechanism
We no longer need to try to sniff the gic version by attempting to
detect the presence of a gic v2 vs v3 at runtime. The sniff is
problematic on at least one platform since it touches registers that may
not be present.

We had switched a while back to detecting the gic version from the
device tree on the one platform it mattered anyway (qemu).

Tested: compile and test on qemu and other arm boards, using both gic
versions.

Change-Id: Ibd86f45e9310fd2709f02086920a4311eb03f690