[fx] introduce fx-error and fx-warn with colors
fx-error prints ERROR: and it's arguemnts to stderr, with the prefix colored
red if stderr is a terminal.
fx-warn prints WARNING: and it's arguemnts to stderr, with the prefix colored
yellow if stderr is a terminal.

test -t is known to be imperfect, however, it should be sufficiently portable
to be usable for now without otherwise going off the deep end.

This should help to make errors and warnings stand out in output for a large
group of users.

Change-Id: Ib7fdee3ca677bfd27e4a3a64d39ecc9a812bfeab