[mediaplayer] Add tests to ReaderCache.
Testing prompted some changes:

1. ReaderCache now returns failure results from the upstream reader
   instead of aborting the process.
2. ReaderCache now supports multiple-load requests from the upstream
   reader instead of crashing. In the future we may want to skip
   the cache entirely for these.

This also caught MTWN-240, which was caused by sliding the buffer
larger distances than its capacity and as a result writing to an
incorrect address. Debug asserts caught this but in release mode
it resulted in corrupt data making its way into the demux.

MTWN-240 #done

TEST: fx run-test mediaplayer_tests -t mediaplayer_demux_tests
Change-Id: Ib0eedc45715ff34c9229381958a3afdd870e3300