[x86][hypervisor] Save and restore extended registers
Zircon saves and restores extended register state on context switching
but not when moving between userspace and the kernel. Therefore, the
hypervisor needs to save and restore on exit and resume to avoid
collisions with the VMM userspace.

Additionally, Zircon does not save all possible extended register state
so for the sake of completeness and safety the hypervisor should save
what the guest is configured to use by setting XCR0 before save/restore.
The value we used is based on the guest's XCR0 and CR4 in the following
way:

x87 - From guest XCR0 (must always be set to 1.)
SSE - Can be enable independently of XCR0, so check CR4.OSFXSR.
AVX - From guest XCR0 (XSAVE-enabled.)
MPX - From guest XCR0 (XSAVE-enabled.)
AVX-512 - From guest XCR0 (XSAVE-enabled.)
PKRU - Can be enable independently of XCR0, so check CR4.PKE.

There are also two pieces of state (PKRU and HDC) that saved/restored
based on IA32_XSS_MSR but the hypervisor does not support that MSR
nor the MSRs that are saved and restored so we never handle those
states.

TEST=Added test to hypervisor-test

Change-Id: I99d4591f4f29408e07d189671d76e9cb2d55b7cf