[inspect] Refactor contained object to be a variant.
We will support both VMO and the current ExposedObject FIDL interface
behind this facade for the time being. This change converts the
fit::optional type (denoting there may or may not be a
component::Object) to fit::variant (denoting that the object may be one
of several types, including empty). The same process was used to convert
StaticMetric and Property types.

A follow up will add a new variant for the VMO types.

TEST=runtests
CF-217: #progress

Change-Id: I6f80434673272e8f883784bbb08002556eca7de2