[debugger] Actually kill processes.
Previously "kill" was accidentally hooked up to "detach". On top of
that, all of the kill messaging (if it were to be called correctly)
still reported "detach".

This patch hooks up the kill command and the termination reasons. Fixes
the zx_status_t type usage in process startup and termination.

Removes the debugged process right away from the debug agent in response
to a Kill command. This matches the Detach usage, and prevents thread
notifications from being sent to the client for a process it's not
attached to any more.

Change-Id: I0f79d8880d35a6bc6e30dc577278dfdb46e1ca14