[devhost] Avoid spurious warnings when device_add fails
These branches did not use the fail() wrapper, which puts flags on the
device indicating it is safe to destroy.  Without it, the logs get a
scary sounding message:
"device: 0x447c486a6540(first_child): only mostly dead (this is bad)"

To make this handling more robust, we switch to an autocall for adding
the flags.

Change-Id: I0d85eed3b266aea0c5ed6bba7b989453c7b3827c