[cobalt] Refactor the testapp so that it doesn't send so often.
The Clearcut server imposes throttling: it is not possible to send to it
more than about once every 5 to 10 seconds. The previous organization of
the testapp made use of functions of the form LogFooAndSend() which
performed both a log and a send.

Recently some changes were made to the testapp which caused it to invoke
these LogFooAndSend() methods hundreds of times. This is a problem because
each invocation of LogFooAndSend() takes 5 to 10 seconds.

In this CL we introduce new functions of the form LogFoo() (without
the "AndSend"). We refactor the tests to invoke LogFoo() rather than
LogFooAndSend(), and then invoke Send() at the end. So we are still
logging the same thing but we are not sending between each log.

Change-Id: I94ff2f51a53cbe25865dff4d8edd26a612f6f9d0