[debugger] Don't take OutputBuffer by copy.
For some reason that's no longer valid I thought Console::Output would
be more efficient if an OutputBuffer could be moved into it. It doesn't
take ownership so there's no benefit.

The result is unnecessary copies if you forget to move, and lots of ugly
move calls. This removes the moves and passes by const ref.

Change-Id: Iad76d8f011c6671fd2c47ddca04f1c14a3c38d77