[prebuilt_package] Initialize package_set in unreferenced packages
This ensures that prebuilt_package() templates which are expanded but
not referenced in any package sets. This can happen if a BUILD.gn rule
defines a test and a production package and the build configuration
depends on all of the tests but doesn't put the production package in
any particular set.

Change-Id: I9d817032a018ae3bae50526d19eeb6db4860e15e